+++ src/main/java/ii/cipriantarlev/marketmanagementapi/documenttype/DocumentTypeRepository.java
@@ -5,6 +5,6 @@ package ii.cipriantarlev.marketmanagementapi.documenttype;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DocumentTypeRepository extends JpaRepository<DocumentType, Integer> {
+public interface DocumentTypeRepository extends JpaRepository<DocumentType, Long> {
 
 }
